+++ app/Http/Controllers/RoleController.php
@@ -137,6 +137,11 @@ class RoleController extends Controller
      */
     public function destroy($id)
     {
+        //Allow admin role to delete only
+        if (!in_array('admin', Auth::user()->roles->pluck('slug')->toArray())):
+            return redirect('/stores')->with('errors', 'Access denied!');
+        endif;  
+        
         $role = Role::find($id);
         if ($role) {
             $role->delete();
